backend/app/services/ai_models.py
"""
AI Model implementations for image processing pipeline.
These are placeholder implementations that you can replace with actual AI models.
"""

import random
from PIL import Image
import time


def apply_image_classification(image_path: str) -> dict:
    """
    Placeholder for image classification model.
    """
    time.sleep(0.3)  # Simulate processing time

    # Mock results
    classes = ["landscape", "portrait", "urban", "nature", "indoor"]
    selected_class = random.choice(classes)

    return {
        "model": "ResNet50",
        "top_prediction": selected_class,
        "confidence": round(random.uniform(0.75, 0.99), 2),
        "top_5": [
            {"class": selected_class, "confidence": 0.89},
            {"class": random.choice(classes), "confidence": 0.06},
            {"class": random.choice(classes), "confidence": 0.03},
        ],
    }


def apply_denoising(input_path: str, output_path: str) -> dict:
    """
    Placeholder for image denoising.
    """
    time.sleep(0.4)  # Simulate processing time

    try:
        # Open image
        img = Image.open(input_path)

        # Save processed image
        img.save(output_path)

        return {
            "model": "Neural Style Transfer",
            "style_applied": "artistic_edge_enhancement",
            "output_path": output_path,
            "success": True,
        }
    except Exception as e:
        return {"model": "Neural Style Transfer", "success": False, "error": str(e)}


def apply_captionning(image_path: str) -> dict:
    """
    Placeholder for image captioning model.
    """
    time.sleep(0.2)  # Simulate processing time

    try:
        img = Image.open(image_path)

        # Get image dimensions
        width, height = img.size

        return {
            "model": "Canny Edge Detection",
            "edges_detected": random.randint(1000, 5000),
            "image_dimensions": {"width": width, "height": height},
            "edge_density": round(random.uniform(0.1, 0.4), 2),
        }
    except Exception as e:
        return {"model": "Canny Edge Detection", "success": False, "error": str(e)}
