backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard-no-fastapi-cloud-cli]>=0.119.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.44",
]
